# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bizosto ERP API
  version: 1.0.0
  description: >-
    Multi-tenant API contract generated from App Router route handlers.
servers:
  - url: /
security:
  - BearerAuth: []
...
paths:
...
  /api/signup:
    post:
      summary: POST /api/signup
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
    put:
      summary: PUT /api/signup
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT